More notes in how to model and a svg fix
I hadn't printed the actual key yet and realized that it was too wide by a new microns I successfully got it in by filing it down but wanted  to update it here. Note to do so I just modified the warding path in the svg using inkscape saved it and reran make clean and make.
I also updated X103-KW12.scad so the parse.py file would accept it. note it needs you to set a variable called name and variables called outlines_k and wardings_k (can't do them all in a row I don't really know why they made it that way)