Added more reference material and fixed schlage.svg for generating
By using an extension for inkscape called apply transforms I updated the schalge.svg file itself so when regenerated in the future it will generate the .gen.scad file with all the shapes in the correct orientation.

If this sort of issue comes again compare the .gen.scad file (specifically uncommenting the outling polygon at the bottom) to a know good one. make sure that the key looks to be in the same orientation to the known good one.